Support html attributes on RootElement
Couldn't get the divs out of the document.

At least now we can set attributes on them.

https://github.com/redfin/react-server/issues/13
