+++ packages/react-server/core/ClientController.js
@@ -12,6 +12,7 @@ var React = require('react'),
 	PageUtil = require("./util/PageUtil"),
 	TritonAgent = require('./TritonAgent'),
 	FramebackController = require('./FramebackController'),
+	{getRootElementAttributes} = require('./components/RootElement'),
 	{PAGE_LINK_NODE_ID, PAGE_CONTAINER_NODE_ID} = require('./constants');
 
 var _ = {
@@ -385,6 +386,11 @@ class ClientController extends EventEmitter {
 
 			element = React.cloneElement(element, { context: this.context });
 			React.render(element, root);
+
+			_.forEach(
+				getRootElementAttributes(element),
+				(v, k) => root.setAttribute(k, v)
+			);
 
 			totalRenderTime += timer.stop();
